Move tabs into engine_utils
